requirements.txt
# Main requirements file that combines all dependencies
-r requirements-core.txt
-r requirements-ml.txt
-r requirements-django.txt

# Core Dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.2
pydantic>=1.9.0,<3.0.0
sqlalchemy==1.4.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==4.3.4
fhirclient>=3.2.0
durable-rules==2.0.28
prometheus-client==0.12.0
python-json-logger>=2.0.7
psutil==5.9.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
requests>=2.26.0
httpx>=0.19.0
tenacity==8.0.1
pyjwt>=2.1.0
aiohttp==3.8.1
pyyaml==6.0
cachetools>=5.3.2
structlog==21.5.0
sentry-sdk==1.5.8

# Machine Learning & AI
numpy==1.24.3
scikit-learn==1.0.2
shap==0.40.0
transformers==4.35.2
torch==2.1.1
openai==1.3.0

# Rules Engine
durable-rules==2.0.28

# Monitoring & Logging
prometheus-client==0.12.0
prometheus-fastapi-instrumentator>=6.1.0
python-json-logger>=2.0.7
psutil==5.9.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.21.1

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
requests>=2.26.0
httpx>=0.19.0
tenacity==8.0.1
pyjwt>=2.1.0
aiohttp==3.8.1
pyyaml==6.0
cachetools>=5.3.2
structlog==21.5.0
sentry-sdk==1.5.8

# Django and related packages
django-redis==5.2.0
django-prometheus==2.2.0
django-health-check==3.16.5
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
django-storages==1.12.3
django-celery-beat==2.2.0
django-celery-results==2.2.0
celery==5.2.3
flower==1.2.0
gunicorn==20.1.0
whitenoise==6.2.0
django-debug-toolbar==3.5.0
django-extensions==3.1.5
django-environ==0.8.1
django-cleanup==6.0.0
django-import-export==2.8.0
django-guardian==2.4.0
django-allauth==0.47.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-framework-social-oauth2==1.0.7
django-oauth-toolkit==2.1.0 